Add hot restart integration test
This test demonstrates that it's possible for clients to receive
unexpected ECONNRESET errors during hot restarts.

We skip the test on JRuby since connection reset errors are expected on
that platform because we cannot reliably pass open file descriptors
(such as the listening sockets) to the new process after a hot restart.