Shutdown client sockets before close
Executing `close` on a tcp socket on a POSIX system doesn't necessarily
flush all writes that the kernel has acknowledged, but not yet sent out.
Executing `shutdown` before `close` communicates to the kernel that the
server is finished writing to the socket and blocks until all pending
writes are sent.